+++ src/app/elements/button/button.component.ts
@@ -10,8 +10,8 @@ export class FormlessButtonComponent {
   @Input() text: string;
   @Input() disabled: boolean = false;
   @Input() color: string;
-  @Input() height: number;
   @Input() inverse: boolean = false;
+  @Input() inputStyles: string;
 
   constructor() {}